(GH-2311) Initial LXD transport work
This adds the basic structure for a new LXD transport that will ship
with Bolt. The transport has one configuration option, `remote`, which
accepts a remote LXD host to connect to. This includes a new VM in the
Vagrantfile that's used for testing. The transport is largely modeled
off of the Docker transport, since they behave relatively similarly by
shelling out to localhost to interact with the containers they are
running.

!feature

* **Add LXD transport** ([#2311](https://github.com/puppetlabs/bolt/issues/2311))

  **This feature is experimental**. Bolt now includes a new LXD
  transport to use when connecting with containers managed with LXD. See
  https://puppet.com/docs/bolt/latest/bolt_transports_reference.html for
  configuration options.