(GH-2314) Fix module name regex for Forge/Git specifications
This updates the regex used to validate module names for Forge and Git
specifications. Previously, the regex for validating the owner segment
of the module name was `[a-z][a-z0-9_]`, which is the same regex used
for module name segments. However, owner names follow different
requirements and can only include letters (uppercase and lowercase) and
digits.

!bug

* **Fix module name validation for Forge and Git module specifications**
  ([#2314](https://github.com/puppetlabs/bolt/issues/2314))

  Forge and Git module specifications now correctly validate the
  module's name and permit uppercase letters in the owner segment of the
  module name. Previously, if the owner segment of a module name
  included uppercase letters, Bolt would raise an error.