(bug) Allow Bolt to use Puppet 6.19.z releases
Previously we had pinned Bolt to depend on Puppet >= 6.18.0 <=6.19.0
because when Puppet released 6.19.0 we picked it up before we had made
Bolt-side changes necessary to be compatible with the new Puppet
version. We restricted the Puppet gem to not pick up new versions so
that we could actively make similar updates in the future without
needing to do so immediately. However, this meant that some gems which
depend on Bolt and Puppet ended up resolving to an old version of Bolt,
which broke their use of the gem. This change means we'll pick up Z
releases of Puppet, but can still make any changes in Bolt we need to
before picking up the Y release.

!bug

* **Allow Puppet 6.19.1 as a dependency**

  This let's Bolt use Puppet 6.19.1 as a dependency.