(GH-2140) Document escalating privilege in powershell scripts on Windows
The WinRM transport does not - and will not - support something akin to
a `run-as` command. This is because the way WinRM works is by connecting
with the user that has permissions to perform the actions, not by
elevating or changing to another user after connection. Put another way,
WinRM operates with a session level token and has no functionality to
change to another user after authentication.

Instead of escalating the user once connected, users can pass
credentials to scripts or plans running on the targets and use them
directly in commands. There are a number of ways to do this, each ideal
for different use cases. This adds a section to the 'Escalating
Privilege' document that describes the different ways to pass
credentials securely to either a script or plan to be used for
escalating on Windows. This includes using sensitive task parameters,
and describes setting and retrieving variables on Target objects
securely using Hiera data, Inventory plugins, and Remote tasks.

Closes #2140

!no-release-notes