(GH-2303) Add module-install config option
This adds a new `module-install` config option that is meant to replace
`puppetfile`. The config option accepts the same sub-options as
`puppetfile` does.

Currently, this config option is not respected when resolving module
dependencies, as changes need to be made to the `puppetfile-resolver`
library to support proxies and an alternate Forge. However, this config
is still passed to r10k, which will respect it when installing modules
from a Puppetfile.

If a user has `module-install` configured and then runs `bolt module
add|install` or `Add|Install-BoltModule`, Bolt will issue a warning
letting the user know that the config option will not be respected when
resolving modules, and then provides a link to a known issue with more
information.

This also adds a handful of warnings that are issued depending on
whether `puppetfile` or `module-install` are configured and if the user
runs `puppetfile install` or `module install`. If both options are
configured, Bolt will also warn and use `module-install`.

This also updates the project migrator to update the `puppetfile` option
to `module-install`.

!feature

* **Add `module-install` configuration option**
  ([#2303](https://github.com/puppetlabs/bolt/issues/2303))

  Bolt now supports a `module-install` configuration option in
  `bolt-project.yaml` and `bolt-defaults.yaml`. This option is used to
  configure proxies and an alternate forge when installing modules using
  the `bolt module add|install` commands or `Add|Install-BoltModule`
  cmdlets.

  _This option is not currently supported when resolving module
  dependencies._