(feature) Add bolt plan new --pp flag
This adds the flag `--pp` to the `bolt plan new` command, which
instructs the command to generate a new Puppet language plan instead of
a YAML plans. The YAML plan is still default, and the plans themselves
are equivalent. This additionally move the `plan new` handler from the
CLI class into it's own `PlanCreator` class in order to separate
concerns.

!feature

* **'bolt plan new' now accepts '--pp' flag to generate Puppet plan**

  The 'bolt plan new' command now accepts a '--pp' flag to generate
  a Puppet plan.