(GH-2304) Set default ports for Puppet services if not specified
Previously, the docs contained a few examples of the `service-url`
config needed to connect Bolt to the Orchestrator that didn't include
the port. Users assumed this meant the port didn't need to be set in
config, and tried to connect to the orchestrator using the default HTTPS
port 443 instead of the typical Orchestrator port 8143. This updates the
documentation to include the port for URLs that don't have standard
ports (80 for HTTP, 443 for HTTPS). It also sets the default ports for
both PuppetDB URLs and Orchestrator URLs to 8081 and 8143 respectively
if the user does not specify a port.

!feature

* **Set default ports for PuppetDB and Orchestrator** ([2304](https://github.com/puppetlabs/bolt/issues/2304))

  We now set the ports for PuppetDB `server_urls` and Orchestrator
  `service-url` to 8081 and 8143 respectively if the port is not set in
  config.