(GH-2302) Send analytics about how many plan types exist in each project
This adds two new GA dimensions to the data we send for user screenviews
when Bolt is run on the CLI: puppet plan count and yaml plan count.
These dimensions simply count the number of files in the project's
`plans` directory (including subdirs) to provide analytics for how many
project-level plans in each language the user has. The dimension is a
user-level metric, meaning that whenever the values for a particular
user change they will get updated in GA and the old value will be
removed. This poses a problem for users with multiple projects, as each
time they run Bolt from a different project their values will get
updated. However the tradeoff is that we get per-project values as
opposed to aggregated values, and we don't get multiple values when a
user adds a new plan.

Closes #2302

!no-release-note