(maint) Update rubocop to 1.3.0
Rubocop released 1.3.0, this adds new cops to our rubocop config and
updates code to adhere to those cops. This includes just three cops:
Duplicate code branches (so if or case statements where the body of the
statement is the same for multiple conditions) which resulted in all of
the code changes, as well as a cop to catch empty classes an nil
lambdas.

!no-release-note