(GH-2321) Move plan_hierarchy docs to writing plans doc
This moves our documentation about the `plan_hierarchy` Hiera config key
to the Writing Plans in the Puppet Language document. Users will only
need to know about `plan_hierarchy` when writing plans, and are more
likely to look for how to use Hiera data in plans (particularly outside
of apply blocks) in the document that isn't about apply blocks. This
also updates the example to be more expressive and close(ish) to a real
world use case.

This additionally standardizes uses of "lookup" to "look up", unless
referencing the plan function "lookup()".

Closes #2321

!no-release-note