(GH-2346) Use target name as Puppet certname
In the Bolt inventory, the Target object's host can be set either by
specifying `uri` or by specifying `host` under the transport
configuration for that target. This means that Target's will not always
have a `uri` specified. Target's will always have a `name` specified,
which is typically how users refer to the target anyway. This updates
the query_resource, custom_fact, and apply_catalog tasks we use when
applying Puppet catalogs or querying Puppet resources to set the Target
`name` as the Puppet certname instead of the `uri`. This is so that
Target's without a URI can use these tasks without failing, and also so
that certname is a bit more predictable for users.

Closes #2346

!bug

* **Targets without a uri can now use apply() and get_resources()** ([#2346](https://github.com/puppetlabs/bolt/issues/2346))

  Previously, if a target had a `host` set instead of a `uri` it would
  error when trying to set the Puppet certname to the target's URI. We now
  use the target's `name` instead of the `uri` as the Puppet certname when
  compiling catalogs.