(GH-2355) Do not error when Psych errors do not have a line number
This updates the `Bolt::Util.read_yaml_hash` method to not error with a
stacktrace when handling Psych errors that do not include a line number.
Previously, if a Psych error did not include a line number, such as the
`Psych::DisallowedClass` error, Bolt would raise an undefined method
error with a stacktrace.

!no-release-note