(GH-2338) Remove maximum version for Puppet gem
We will continue to ship Bolt with the puppet 6 gem, but we shouldn't
need to limit the maximum version of the Puppet gem for now. It seems
like Bolt is compatible with Puppet 7, and we don't have any breaking
changes going in to Puppet that we need to gate against. This also means
tools that use Bolt as a dependency won't be limited to using Puppet 6.x.

!no-release-note