(GH-2362) Deprecate boltdir, configfile, puppetfile, and description CLI flags
This adds deprecation warnings for the `--boltdir`, `--configfile`,
`--puppetfile`, and `--description` command line flags. Whenever users
use any of these command line options - and only if the option is
specified on the CLI - a warning should be printed. This doesn't update
spec tests or other internal references to deprecated options, but it
does update the help text to not include deprecated flags.

Closes #2362

!deprecation

* **Add deprecation warnings for boltdir, configfile, puppetfile, and description** ([#2362](https://github.com/puppetlabs/bolt/issues/2362))

  We are planning to remove the `--boltdir`, `--configfile`,
  `--puppetfile`, and `--description` command line flags in the next major
  version of Bolt. This adds deprecation warnings that are printed when
  users specify any of these flags.