(GH-2313) Add plugin caching
This adds some basic caching for Bolt plugins. Caching is enabled by
configuring `cache`, either as a default cache for all plugins in the
`plugin-cache` config option, or as configuration on the plugin
itself with the `_cache` key. When the plugin reference is evaluated, if
caching is enabled Bolt will check the cache first to see if the plugin
result has been cached and is fresh.

Bolt stores cached results in the `.plugin_cache.json` file of the
bolt project and assigning each plugin result a unique ID by hashing a
stringified, sorted version of the hash. This means that if users have
the exact same plugin with the same configuration in multiple places in
the inventory they can share the same cache files. Bolt will check the
cache and see if cache is past it's ttl. If it's not, Bolt will first
clear out all outdated cache results from the cache file before
returning the cached result for this plugin. This ensures that orphaned
cache entries (results whose plugins have been changed or removed) don't
stay forever. Bolt is able to clear out orphaned results by storing the
original ttl alongside the results in the cache file. This does have the
side effect that a plugin may be cleared even if it's ttl is updated, if
it has already reached it's previous ttl.

If the cache is past it's ttl or if no cache has been configured Bolt
will evaluate the plugin as normal and cache the result. It stores the
current ttl with the result so the results can be cleaned out.

This also provides `--clear-cache` and `-ClearCache` which
clears all cached results.

There are a lot of other features that would be interesting to add:
- Encrypted caching
- Refreshing only certain plugins

Closes #2313

!feature

* **Plugin caching**

  **This feature is experimental**. Bolt plugins can now be configured
  to cache their results. Users can either configure a default cache
  time-to-live for all plugins, or configure each plugins ttl
  individually. See https://pup.pt/bolt-cache for more information