(GH-2376) Warn about unknown configuration options
This updates the config validator to issue warnings for unknown
config options. When the validator encounters an unknown option, it
stores a warning indicating where the unknown option is specified, which
is later issued by the logger. Options are only reported as unknown if
the schema for the option defines the `:properties` key but _not_ the
`:additionalProperties` key.

!feature

* **Warn about unknown configuration options**
  ([#2376](https://github.com/puppetlabs/bolt/issues/2376))

  Bolt now issues a warning when it detects an unknown configuration
  option. The warning will indicate where the configuration option is
  located.