(GH-2365) Deprecate Powershell 2 support
This adds a deprecation warning when users run Bolt from Powershell
version less than 3, and for each target users run against in Powershell
less than 3. This only applies to the shell Bolt is run from or
connected to, not all Powershell versions on the machine. A warning is
issued for each target the user connects to with a deprecated version so
that users know which targets have deprecated PS versions.

Closes #2365

!deprecation

* **Deprecate Powershell 2 support** ([2365](https://github.com/puppetlabs/bolt/issues/2365))

  Support for Powershell 2 on both Bolt targets and controllers is
  deprecated, and will be dropped in Bolt 3.0.