(GH-2379) Suggest noexec may be an issue with exit code 126
Exit code 126 indicates that a file Bolt has written to the temporary
directory can't be executed because the user Bolt is running as doesn't
have permission to execute it. This can be caused by a variety of
issues, but one common cause is that administrators have mounted the
default tmpdir with `noexec`, meaning that users can't execute files in
the tmpdir. The recommended workaround for this is to configure an
alternative `tmpdir` for the transport(s) being used, or to contact the
administrator about modifying the permissions (though usually they're in
place for a reason). This appends a brief message that `noexec` may be
the issue to stderr output whenever exit code 126 is returned. This is
appended for bash shells. The message directs users
to our troubleshooting doc, which has been updated to include a note
about `noexec` and it's mitigation.

Closes #2379

!no-release-note