(bug) Add 'notice' log level to schema and deprecate
This adds the `notice` log level back to the config schema, which was
removed in #2098. Because we now use the schema for config validation,
the absence of `notice` as a valid option resulted in Bolt raising an
error for any user that had that log level configured. Now, Bolt will
instead issue a deprecation warning that the log level is deprecated and
will be removed in 3.0.

!bug

* **Support `notice` log level**

  Log levels can now be set to `notice`. Previously, Bolt would raise an
  error saying that `notice` was not a supported log level.

!deprecation

* **Deprecate `notice` log level**

  The `notice` log level is deprecated and will be removed in Bolt 3.0.
  Use the `info` log level instead.