(maint) Remove 'remote' and 'workstation' from running commands page
The 'Running Bolt commands' page often refers to "remote targets" when
in fact the documentation also applies to using local targets. It also
uses 'workstation' to refer to the machine Bolt is run from, when "Bolt
controller" is a more accurate and clear term. This removes uses of
`remote` where applicable (with the exception of authenticating
connections, which is not relevant to the local transport), and updates
'workstation' to 'Bolt controller'.

!no-release-note