Add puppet_connect_data plugin for bolt_server
This adds a basic implementation of the puppet_connect_data plugin that
takes a data hash and looks up keys in the hash. The data is expected to
be supplied as a map in the following structure:

    {
      "key1": {
        "value": "value1"
      },
      "key2": {
        "value": "value2"
      }
    }

The plugin will fail if no value is supplied for the key. It
specifically does *not* validate that each value map has a 'value' key,
as that validation should be done prior to passing the value to the
plugin constructor.

Currently, this plugin is not loaded anywhere.