(GH-2400) Add bundled-ruby option for local transport
This introduces the `bundled-ruby` config option for the local
transport. The option enables merging default config values currently
used just for a target with name `localhost`. Those default config
values are:

```
config:
  local:
    interpreters:
      '.rb' => RbConfig.ruby # This is Bolt's bundled Ruby
features:
  - puppet-agent
```

The default config values are merged at the target config level, meaning
they are overridden by target-level config but not group-level config.
The 'features' array is appended, not overridden, and is a set so the
value will not be added twice.  The `localhost` target additionally
defaults to using the `local` transport as part of the "magic config",
however the `bundled-ruby` option will only apply to targets that are
already configured to use the local transport.

`bundled-ruby` currently defaults to false, but will default to `true`
in Bolt 3.0. A warning is issued for users who don't set the
conifguration and use the local transport that the default config will
be updated.

Closes #2400

!feature

* **bundled-ruby local transport config option enables local defaults** ((#2400)[https://github.com/puppetlabs/bolt/issues/2400)

  Set `bundled-ruby` in the local transport config to enable or disable
  the default config currently used for the 'localhost' target.