(GH-2417) Update getting started to use module workflow
The new module workflow is now stable, and as part of moving it out of
experimental we now add a `modules` key to `bolt-project.yaml` when
users run `bolt project init` or `New-BoltProject`. Our getting started
guide instructs users to use the project init command to create their
project, which enables the new workflow. As part of the new module
workflow the `site-modules/` directory is not on the modulepath by
default, and instead we expect users to use the `modules/` directory for
any custom modules they are developing. This updates the getting started
guide to instruct users to create their new module inside the `modules/`
directory within their project instead of in the `site-modules/`
directory.

Closes #2417

!no-release-note