(feature) Print spinner to the console for long operations
This introducs a spinner which spins while Bolt performs long
operations. Currently the spinner is added for installing modules,
running plans, and running tasks.

The spinner works by creating a new thread and continuously printing a
series of messages, where each message includes a backspace and the
next character from an array of characters. The spinner is a function of
the outputters, which each have their own `start_spin` and `stop_spin`
functions that open and close the spinner thread respectively.

The spinner is enabled by default, and can be disabled by setting
`spinner: false` in any of `bolt.yaml`, `bolt-project.yaml`, or
`bolt-default.yaml`.

!feature

* **CLI spinner for long running operations**

  Bolt now has a spinner printed to the CLI for long-running operations,
  so that users know the Bolt process has not hung. Disable the spinner
  by setting `spinner: false` in any Bolt configuration file.