(GH-2413) Validate inventory with validator
This updates the `Bolt::Inventory` class to validate loaded inventory
against a schema using the validator. If the loaded inventory has any
errors, Bolt will raise an error describing each error, as well as its
location.

This also updates the `Bolt::Config::Validator` class. First, the class
is now just `Bolt::Validator`, since it is used to validate more than
just config files. Additionally, the validator now expects a schema
describing the _entire_ data structure that was loaded, not just the
various options available. This means each schema needs to describe the
type of data passed in. Because of this change, each of the schema
methods on `Bolt::Config` and `Bolt::Project` have been updated to
describe the full data structure.

To accommodate nested `groups` in inventory, the validator now handles a
new `:_ref` key in the schema. When the validator comes across a `:_ref`
key, it will use a definition from the schema to validate the value
against, pulling the definition from `schema[:definitions][ref]`. This
allows the validator to perform recursive validation.

!feature

* **Improved inventory validation**
  ([#2413](https://github.com/puppetlabs/bolt/issues/2413))

  Bolt now validates inventory against Bolt's inventory schema and
  indicates where errors are found.