(GH-1549) Add private "metadata" to plans
This adds a `private` parameter to plan "metadata" that tells Bolt
content consumers to not display the plan. For YAML plans the parameter
is a top-level `private` key that takes a boolean value. For Puppet
plans the parameter is a puppet strings tag `@private` which accepts the
strings 'true' and 'false'. The `bolt plan convert` command knows how to
translate the `private` key from YAML plans to Puppet plans.

Plans that have the `private` parameter set to true will not be shown in
`bolt plan show` or `Get-BoltPlan` lists. Plan details can be shown with
`bolt plan show <plan>` or `Get-BoltPlan -Name` if the plan name is
given. Private plans can still be run with Bolt.

Bolt will error if private is defined but has an invalid value.

This also adds caching for plan `private` metadata whenever
modules are downloaded or a `generate-types` command is run. If PAL is
running without a project, caching is not enabled. Caching works by
saving the plan details hash that's used for `bolt plan show <plan>` to
`.plan_cache.json` in the Bolt project. The cache is updated whenever a
`generate-types` command is run, when individual "local plans" are updated
if they have been modified since they were cached, or when the user
removes the cache file. The cache stores the modified time of the
local plan when it's cached, and checking it against the current
modified time of the plan when it's going to be listed. A plan is
considered local if it's in the project's `plans/` directory, or in a
module in the `modules/` directory of a project. All other plans are
considered downloaded, and can only be refreshed by re-downloading the
plan or removing the cache file. Only the `private` metadata is cached,
no other data is saved.

Closes #1549

!feature

* **Hide private plans from 'bolt plan show'** ([#1549](https://github.com/puppetlabs/bolt/issues/1549))

  Users can now set the top-level `private` key in YAML plans, or the
  `@private` puppet string.