(GH-2428) Update with_project method to be more transparent
This updates the `BoltSpec::Project` method `with_project` to be more
transparent. Previously, the method used some magic to configure a
project by calling the `project_config` method. Additionally, the method
would also set several instance variables for spec tests to use.

The `with_project` method now includes a few keyword arguments that can
be used to modify the project directory structure that it creates:

- `config`

  A hash of config to write to `bolt-project.yaml`. If `nil`, writes a
  default config file with the project name set.

- `inventory`

  A hash of inventory data to write to `inventory.yaml`. If `nil`,
  writes an empty inventory file.

- `boltdir`

  A boolean. When true, creates the project inside a `Boltdir`.

The `with_project` method now yields a `Bolt::Project` instance for the
project directory. The `in_project` method was also updated to accept
the same keyword arguments and yield a `Bolt::Project` instance.

This also adds two new methods to `BoltSpec::Project`:
`with_project_directory` and `make_project`. The
`with_project_directory` method accepts all the same arguments as
`with_project`, but yields the project path instead of a `Bolt::Project`
instance. This method is helpful when writing tests that rely on errors
being raised when project config is loaded. The `make_project` method
accepts a path and returns a `Bolt::Project` instance for the project
directory.

!no-release-note