(GH-2449) Remove deprecated CLI flags from the codebase
This removes `--configfile`, `--boltdir`, and `--puppetfile` (and
`--description`, but that wasn't used anywhere) from our spec tests. It
leaves tests that specifically test those flags to be removed when the
flag is removed, but refactors tests that relied on the flags for
setting Bolt config to use Bolt projects instead.

!no-release-note