(bug) Allow _cache in PuppetDB plugin options
We validate that the PuppetDB plugin that ships with Bolt only contains
valid keys in the plugin class itself (not through schema validation).
We missed adding `_cache` as a valid key to the PuppetDB plugin options,
which resulted in an error when users try to configure a cache for the
PuppetDB plugin. This adds `_cache` as a valid key for the PuppetDB
plugin.

!bug

* **Allow caching for PuppetDB plugin**

  Previously, our configuration validation would raise an error if users
  supplied `_cache` to the PuppetDB plugin. Cache is now configurable
  for the plugin.