(GH-2429) Remove bolt.yaml from spec tests
This removes `bolt.yaml` and replaces it with `bolt-project.yaml` file
infrastructure for any tests that aren't explicitly testing behavior
around the presence of `bolt.yaml`. This mostly includes refactoring the
plugin integration test to use `with_project` instead of their own
custom functions.

This also edits the `with_project` spec test helper to set the project
name by default, and merge in the provided config over top of `{ 'name'
=> name }` if config is passed to the helper. This ensures the project
always has a name, and that the name can be overridden by the config
hash if need be.

Closes #2429

!no-release-note