(maint) Unpin bundler for Bolt
We have a CI test that verifies that Litmus works with each Bolt PR by
adding a gem declaration to the Gemfile of
[puppetlabs/puppetlabs-package](https://github.com/puppetlabs/puppetlabs-package)
that pulls in the Bolt PR as a gem and runs tests in the module.
Installing gems in the module was hanging in Github Actions when Bundler 2.2.0
was released, due to [this
issue](https://github.com/rubygems/rubygems/issues/4123). We originally
pinned all of Bolt back to bundler < 2.2.0, however now the Jenkins CI
environment has been updated to only include Bundler 2.2.1. In order to
be able to test and build Bolt in Jenkins we need to unpin the bundler
version so we can use 2.2.1.

The original issue has also been fixed now, so we can unpin bundler
entirely.

!no-release-note