(maint) Enable Rubocop 1.7.0 cops
This enables cops that were added to Rubocop 1.7.0. Enabling the cops
didn't result in any new errors, so just the configuration file change
is needed.

There was a separate issue where an integration test verifying the
Puppetfile resolver error when dependencies fail to resolve now fails
for a different module than the one the test checks. The failure is
still as expected, so this updates the check to just check for the error
message and not the module name.

!no-release-note