(maint) Refactor the Docker transport to use Shell class
This refactors the Docker transport to use the same codepaths that
other transports do, in an effort to prove viability of making
transports pluggable. The behavior of the transport should not change at
all.

The `Bolt::Transport::Docker` class now inherits from the `Simple`
transport class, and only overloads the `with_connection` function to
create a new `Bolt::Transport::Docker::Connection` instance and yield
it.

The `Bolt::Transport::Docker::Connection` class provides the functions
that the shell classes need to interact with Docker containers:
`connect`, `upload_file`, `download_file`, and `execute`. These are the
functions we're hoping are needed for a stable API for making transports
pluggable. This update largely consists of renaming functions to match
the API, but also includes distinguising the `execute` function that the
shell classes uses that must return pipes, and the
`execute_local_docker` functions that are used internally by the
Connection class and just return strings.

!no-release-note