(maint) Handle YAML plan errors when transpiling
Previously, trying to convert a YAML plan with type or syntax errors
would result in a stacktrace, as the validation error raised by the YAML
plan loader would not be handled. When loading YAML plans to show or run
in `Bolt::PAL` we load the plans using the `in_bolt_compiler` block,
which catches such errors and raises a Bolt error. We don't transpile
plans using `in_bolt_compiler` since we don't need to do anything with the
plan other than read it, so don't need a temporary environment or Puppet
settings that `in_bolt_compiler` handles. This is a long way of saying:
we don't need this handling in other places that we load YAML plans as
errors are already handled by `in_bolt_compiler`, but we do need it in
the transpiler. The transpiler now rescues possible errors and raises a
Bolt error, with file details if they are available.

!bug

* **Don't stacktrace when converting YAML plans with errors**

  Bolt will now error cleanly instead of stacktracing when users try to
  convert a YAML plan that has type or syntax errors.