(GH-2517) Support project-level plugins
This change adds support for project-level plugins. When Bolt loads
plugins, it now checks whether the project is a named project and is
loaded as a module. If the project is named, and contains a
`bolt_plugin.json` file, it is added as a plugin module in Bolt. All
plugin hooks are supported for project-level plugins.

!feature

* **Support for project-level plugins**
  ([#2517](https://github.com/puppetlabs/bolt/issues/2517))

  Bolt now supports project-level plugins. Similar to module plugins,
  project-level plugins are implemented as tasks that use specific hooks
  and are referred to using the name of the project.