(GH-2509) Improve PowerShell Task Error Formatting
This commit formats the exception written to stderr so that the line and char position is written to bolt. This position is the actual location inside the script that the bolt task is executing, instead of a relative location inside the code block we 'send' over with all of our helper functions.

Note that this uses the `$host.UI` to 'write' the message we want. This works both in our winrm and local transports because both have a PSHost implementation. This could fail if somehow we don't have access to a `$host`, but this is not a common use case with our transports.

This also formats the message shown in a different order than what is shown normally. It prioritizes the exception type and position first, then shows the exception message. This may not be the best presentation of this information, so we may need to adjust this for different exception types in the future.

!feature

* **Improve bolt powershell task error message** ([#2509](https://github.com/puppetlabs/bolt/issues/2509))

  Format the exception powershell type tasks throw to make it easier for a user to read the error message.