(docs) List which config options are pluggable
This updates the `Using Plugins` document to tell users that not all
configuration options are pluggable, and how to request a new option be
made pluggable. This also updates the configuration reference templates
to add a `Pluggable` list item for configuration options that are
pluggable. The `Using Plugins` page points to this reference for users
that want to see which options are pluggable.

!no-release-note