(GH-2518) Add read-timeout config for PCP transport
This adds a `read-timeout` configuration setting for the PCP transport,
which accepts an integer and passes the configuration to
`OrchestratorClient.new()`. Although the option is an integer, it
should be nil if a value is not provided, as there's no static default
for `Net::Http::Persistent#read_timeout`, so we should not set one
unless it's configured. This allows users to configure either failing
quickly if the Orchestrator is busy (such as during code deploys), or to
extend the time if they'd like to wait for the jobs to finish.
`read-timeout` configures the length of time to wait for a response from
the Orchestrator before raising a timeout error when making HTTP
requests.

!feature

* **Add read-timeout configuration option for PCP transport** ([#2518](https://github.com/puppetlabs/bolt/issues/2518))

  Users can now configure a `read-timeout` for HTTP requests to the
  Orchestrator, which defines how long to wait for a response before
  raising a Timeout error.